Move feature modules into features package
