+++ pyproject.toml
@@ -52,16 +52,23 @@ packages = [
   "zapzap.core.i18n",
   "zapzap.core.platform",
   "zapzap.core.theme",
+  "zapzap.features",
+  "zapzap.features.alerts",
+  "zapzap.features.customizations",
+  "zapzap.features.dictionaries",
+  "zapzap.features.downloads",
+  "zapzap.features.notifications",
+  "zapzap.features.permissions",
+  "zapzap.features.startup",
+  "zapzap.features.tray",
   "zapzap.models",
   "zapzap.models.settings",
   "zapzap.resources",
-  "zapzap.services",
   "zapzap.views",
   "zapzap.views.settings",
   "zapzap.views.browser",
   "zapzap.views.components",
   "zapzap.views.settings_components",
-  "zapzap.notifications",
   "zapzap.webengine"
 ]
