+++ zapzap/controllers/DownloadDialog.py
@@ -17,7 +17,7 @@ from gettext import gettext as _
 import os
 
 from zapzap.core.config.SettingsManager import SettingsManager
-from zapzap.services.DownloadNamingService import DownloadNamingService
+from zapzap.features.downloads.DownloadNamingService import DownloadNamingService
 
 
 class DownloadDialog(QDialog):
