+++ zapzap/controllers/main_window_controller.py
@@ -11,9 +11,9 @@ from zapzap.controllers.QtoasterDonation import QtoasterDonation
 from zapzap.controllers.ShortcutsDialog import ShortcutsDialog
 from zapzap.controllers.browser_controller import BrowserController
 from zapzap.controllers.settings_controller import SettingsController
-from zapzap.services.AlertManager import AlertManager
+from zapzap.features.alerts.AlertManager import AlertManager
 from zapzap.core.config.SettingsManager import SettingsManager
-from zapzap.services.SysTrayManager import SysTrayManager
+from zapzap.features.tray.SysTrayManager import SysTrayManager
 from zapzap.core.theme.ThemeManager import ThemeManager
 from zapzap.views import MainWindowView
 from gettext import gettext as _
