+++ zapzap/controllers/settings/advanced_customizations_settings_controller.py
@@ -19,8 +19,8 @@ from PyQt6.QtWidgets import (
     QVBoxLayout,
 )
 
-from zapzap.services.CssPreviewService import CssPreviewService
-from zapzap.services.CustomizationsManager import CustomizationsManager
+from zapzap.features.customizations.CssPreviewService import CssPreviewService
+from zapzap.features.customizations.CustomizationsManager import CustomizationsManager
 from zapzap.core.config.SettingsManager import SettingsManager
 
 from zapzap.views.settings import AdvancedCustomizationsSettingsView
