+++ zapzap/core/environment/SetupManager.py
@@ -1,7 +1,7 @@
 from os import environ, getenv
 from PyQt6.QtCore import QFileInfo
 from zapzap.core.platform import IS_WINDOWS
-from zapzap.services.DictionariesManager import DictionariesManager
+from zapzap.features.dictionaries.DictionariesManager import DictionariesManager
 from zapzap.core.config.SettingsManager import SettingsManager
