+++ zapzap/features/customizations/AddonsManager.py
