+++ zapzap/features/customizations/ExtensionRuntimeManager.py
@@ -1,4 +1,4 @@
-from zapzap.services.CustomizationsManager import CustomizationsManager
+from zapzap.features.customizations.CustomizationsManager import CustomizationsManager
 
 
 class ExtensionRuntimeManager:
