+++ zapzap/features/notifications/__init__.py
@@ -1,0 +1,1 @@
+"""Feature package."""
