+++ zapzap/features/tray/SysTrayManager.py
