+++ zapzap/models/settings/appearance_settings_model.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from zapzap.resources.CSRButtonThemeProvider import CSRButtonThemeProvider
 from zapzap.resources.TrayIcon import TrayIcon
 from zapzap.core.config.SettingsManager import SettingsManager
-from zapzap.services.SysTrayManager import SysTrayManager
+from zapzap.features.tray.SysTrayManager import SysTrayManager
 from zapzap.core.theme.ThemeManager import ThemeManager
