+++ zapzap/models/settings/language_downloads_settings_model.py
@@ -1,8 +1,8 @@
 """Data access for the general settings page."""
 
-from zapzap.services.AutostartManager import AutostartManager
-from zapzap.services.DictionariesManager import DictionariesManager
-from zapzap.services.DownloadManager import DownloadManager
+from zapzap.features.startup.AutostartManager import AutostartManager
+from zapzap.features.dictionaries.DictionariesManager import DictionariesManager
+from zapzap.features.downloads.DownloadManager import DownloadManager
 from zapzap.core.config.SettingsManager import SettingsManager
 from zapzap.core.environment.SetupManager import SetupManager
 from zapzap.core.i18n.TranslationManager import TranslationManager
